+++ BetterSort.LastPlayed.Test/Mocks/InMemoryDateRepository.cs
@@ -5,16 +5,16 @@ namespace BetterSort.LastPlayed.Test.Mocks {
   using System.Linq;
 
   internal class InMemoryDateRepository : IPlayedDateRepository {
-    public Dictionary<string, DateTime> LastPlayedDate { get; set; } = new();
+    public Dictionary<string, DateTime>? LastPlayedDate { get; set; } = new();
 
     public StoredData? Load() {
-      return new StoredData() {
+      return LastPlayedDate == null ? null : new StoredData() {
         Version = "",
         LastPlays = LastPlayedDate,
       };
     }
 
-    public void Save(IDictionary<string, DateTime> playDates) {
+    public void Save(IReadOnlyDictionary<string, DateTime> playDates) {
       LastPlayedDate = playDates.ToDictionary(x => x.Key, x => x.Value);
     }
   }