+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 **Delete previous BetterSongList.LastPlayedSort.dll if you have it.**
 
 - Rename project from BetterSongList.LastPlayedSort to BetterSort.LastPlayed([#1](https://github.com/nanikit/BetterSort.LastPlayed/issues/1)).
+- Import [SongPlayHistory](https://github.com/Shadnix-was-taken/BeatSaber-SongPlayHistoryContinued)'s history data when there is no history (for example first install).
 
 ### 1.0.1
 - Fix not recorded plays sometimes.